pyproject.toml
# Python project configuration
[project]
name = "mcp-server-snowflake"
version = "0.1.0"
description = "MCP server for interacting with Snowflake databases"
readme = "README.md"
requires-python = ">=3.10"  # Required Python version
dependencies = [  # Project dependencies / 项目依赖项
    "mcp>=1.0.0",                    # MCP SDK
    "snowflake-connector-python",     # Snowflake connector
    "python-dotenv"                   # Environment variable management
]

# Build system configuration
[build-system]
requires = ["hatchling"]  # Build tool requirement
build-backend = "hatchling.build"  # Build backend

# Build configuration
[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_snowflake"]  # Package location

# Entry points
[project.scripts]
mcp-server-snowflake = "mcp_server_snowflake.server:main"  # Command line entry